pulls out preferences list fro admin guest into a separate component
